Add files via upload
Client is a meal delivery company that operates in multiple cities. They have various fulfilment centres in these cities for dispatching meal orders to their customers. The client wants to help these centres with demand forecasting for upcoming weeks so that these centres will plan the stock of raw materials accordingly.
The replenishment of most raw materials is done on a weekly basis and since the raw material is perishable, procurement planning is of utmost importance. Secondly, staffing of the centres is also one area wherein accurate demand forecasts are helpful.